requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
sendgrid>=6.11.0
pydantic>=2.0.0
asyncio>=3.4.3
ipython>=8.0.0
